+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".PayPalActivity" android:exported="true" />
+        <activity android:name=".FingerprintActivity" android:exported="true" />
+        <activity android:name=".CCTokenizerActivity" android:exported="true" />
+        <activity android:name=".GooglePayActivity" android:exported="true" />
         <activity
             android:name=".MyCardPaymentActivity"
             android:launchMode="singleTop"
