+++ app/src/main/res/layout/activity_cc_tokenizer.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:padding="24dp">
+
+    <TextView
+        android:id="@+id/cc_tokenizer_title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="CC Tokenizer"
+        android:textSize="24sp"
+        android:textStyle="bold"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginBottom="16dp" />
+
+    <Button
+        android:id="@+id/btnStartCCTokenizer"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Start Credit Card Tokenization"
+        android:layout_marginTop="24dp" />
+
+    <RelativeLayout
+        android:id="@+id/fragment_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
+
+</LinearLayout>
