+++ app/src/main/res/layout/activity_fingerprint.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:padding="24dp">
+
+    <TextView
+        android:id="@+id/fingerprint_title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Fingerprint Tokenizer"
+        android:textSize="24sp"
+        android:textStyle="bold"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginBottom="16dp" />
+
+    <Button
+        android:id="@+id/btnStartFingerprint"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Start Fingerprint Tokenization"
+        android:layout_marginTop="24dp" />
+
+    <TextView
+        android:id="@+id/tvSnippetToken"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Token will appear here"
+        android:textSize="16sp"
+        android:layout_marginTop="24dp" />
+
+
+</LinearLayout>
