+++ app/src/main/res/layout/activity_main.xml
@@ -1,72 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:orientation="vertical"
     android:padding="16dp">
 
     <Button
         android:id="@+id/fingerprint_button"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Start Fingerprint Tokenizer"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/tvTitleFingerprint"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/fingerprint_button"
-        android:text="Snippet token:" />
-
-    <TextView
-        android:id="@+id/tvSnippetToken"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tvTitleFingerprint" />
+        android:text="Start Fingerprint Tokenizer" />
 
     <Button
         android:id="@+id/btnCCTokenizer"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="24dp"
-        android:text="Start CreditCard Tokenizer"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tvSnippetToken" />
+        android:text="Start CreditCard Tokenizer" />
 
 
-    <RelativeLayout
-        android:id="@+id/fragment_container"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintTop_toBottomOf="@id/btnCCTokenizer"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-
-    <androidx.compose.ui.platform.ComposeView
-        android:id="@+id/compose_google_pay"
+    <Button
+        android:id="@+id/btnGooglePay"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/btnCCTokenizer"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="16dp" />
+        android:layout_marginTop="24dp"
+        android:text="Start Google Pay" />
 
-    <com.paypal.android.paymentbuttons.PayPalButton
-        android:id="@+id/paypal_button"
+
+    <Button
+        android:id="@+id/btnPayPal"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/compose_google_pay"
-    />
+        android:layout_marginTop="24dp"
+        android:text="Start Paypal" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
