+++ app/src/main/res/layout/activity_paypal.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:padding="24dp">
+
+    <TextView
+        android:id="@+id/paypal_title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="PayPal"
+        android:textSize="24sp"
+        android:textStyle="bold"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginBottom="16dp" />
+
+    <com.paypal.android.paymentbuttons.PayPalButton
+        android:id="@+id/paypal_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp" />
+
+    <!-- Add more UI elements here as needed -->
+
+</LinearLayout>
